Development

+++ database/migrations/2023_03_22_154030_create_employee_table.php
@@ -72,6 +72,8 @@ return new class extends Migration
      */
     public function down(): void
     {
+      if (DB::getDriverName() !== 'sqlite') {
         Schema::dropIfExists('employee');
+      }
     }
 };
