+++ database/migrations/2023_04_10_092209_create_crm_prospects_table.php
@@ -62,6 +62,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('crm_prospects');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('crm_prospects');
+        }
     }
 };
