+++ database/migrations/2023_04_24_075207_create_directrecruitment_applications_table.php
@@ -35,7 +35,9 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('directrecruitment_applications');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('directrecruitment_applications');
+        }
     }
 };
